+++ labbook_BE/default_settings.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 
-APP_VERSION = '3.2.5'
+APP_VERSION = '3.2.6'
 
 # ######################################
 #  Database parameters